create config module and rework to allow user input
